non-deallocating arena allocator
now that we're using the heap, it seems silly to spend time managing
memory fragmentation and cleanup when the watch is just going to go
to sleep and wipe out ram completely anyway.

instead, use an arena allocator that efficiently allocates and
never deallocates. we'll have to use less than 16kb before the watch
resumes sleeping